requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
RPi.GPIO==0.7.1
mfrc522==0.0.7
sqlite3==2.6.0
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0